Made player credits unsigned long.

+++ source/universe.h
@@ -80,8 +80,10 @@ struct player
   int experience;
   int alignment;
   int turns;
-  int credits;
-  int bank_balance;
+  unsigned long credits;
+  unsigned long bank_balance;
+  //int credits;
+  //int bank_balance;
   int lastprice;		/* Last price offered by a port. */
   int firstprice;		/* first price offered by a port */
   unsigned short intransit;	/* Is the player moving? */
